4 x (Lab Report + Drawing) or total 8 x file uploaded
To prepare lab reports, first relevant drawing is carried out with Wondershare EdrawMax software. Then report is written and attached the drawing using MS Word; and saved as a pdf file. These files are named starting with '1. a)', '2. a)', '3. a)' and so on.

The drawings are also uploaded here named starting with '1. b)', '2. b)', '3. b)' and so on.